Move modal to components, password pages to user
